+++ modules/controllers/01-network.tf
@@ -13,7 +13,7 @@ resource "aws_subnet" "k8s_private" {
 
   tags = {
     Environment                       = var.environment
-    ManagedBy                         = "terraform (k8s module)"
+    ManagedBy                         = "opentofu (k8s module)"
     ModuleRepository                  = "https://github.com/jecnua/terraform-aws-kubernetes"
     Name                              = "${var.unique_identifier} ${var.environment} k8s private subnet"
     KubernetesCluster                 = var.kubernetes_cluster
@@ -30,7 +30,7 @@ resource "aws_subnet" "k8s_public" {
 
   tags = {
     Environment              = var.environment
-    ManagedBy                = "terraform (k8s module)"
+    ManagedBy                = "opentofu (k8s module)"
     ModuleRepository         = "https://github.com/jecnua/terraform-aws-kubernetes"
     Name                     = "${var.unique_identifier} ${var.environment} k8s public subnet"
     KubernetesCluster        = var.kubernetes_cluster
@@ -43,7 +43,7 @@ resource "aws_route_table" "k8s_private_route_table" {
 
   tags = {
     Environment       = var.environment
-    ManagedBy         = "terraform (k8s module)"
+    ManagedBy         = "opentofu (k8s module)"
     ModuleRepository  = "https://github.com/jecnua/terraform-aws-kubernetes"
     Name              = "${var.unique_identifier} ${var.environment} k8s private route table"
     KubernetesCluster = var.kubernetes_cluster
@@ -55,7 +55,7 @@ resource "aws_route_table" "k8s_public_route_table" {
 
   tags = {
     Environment       = var.environment
-    ManagedBy         = "terraform (k8s module)"
+    ManagedBy         = "opentofu (k8s module)"
     ModuleRepository  = "https://github.com/jecnua/terraform-aws-kubernetes"
     Name              = "${var.unique_identifier} ${var.environment} k8s public route table"
     KubernetesCluster = var.kubernetes_cluster