+++ modules/controllers/CHANGELOG.md
@@ -33,7 +33,7 @@ Use the next version in which the correct fix is implemented.
 - Now the cluster is HA with multiple masters :) and THIS was not easy :D
 - You will need to open the security group to the IPs you want to access the cluster from (since it's using a NLB) and the NATs
 - Updated k8s version to 1.27
-- Updated terraform version to <= 1.5.5 (last version I will support. Next one will be openTF)
+- Updated terraform version to <= 1.5.5 (last version I will support. Next one will be opentofu)
 
 ### Features & Changes